refactor(activity-log): extract deleteCompletelyOperation into its own module
Move the complete-deletion primitive out of the 5709-line
service/page/index.ts into service/page/delete-completely-operation.ts.
The PageService method becomes a thin delegator; IPageService and all four
call sites are unchanged. Per the executor principle, the work-set
(pageIds/pagePaths/actor) is passed in and collaborators are injected via
crowi. The prisma comments filter now stringifies page ids explicitly
(String @db.ObjectId expects hex strings), matching the ObjectId-stringify
convention already used by cascade-attachment-removal-inputs; behaviour is
otherwise unchanged.

Verified: tsgo --noEmit clean, biome clean, page-deletion integ green
(cascade/empty-trash/user-home + v5 public/non-public = 74 tests).
