+++ apps/app/src/server/service/page/index.ts
@@ -71,7 +71,6 @@ import { prepareDeleteConfigValuesForCalc } from '~/utils/page-delete-config';
 import { prisma } from '~/utils/prisma';
 
 import type { ObjectIdLike } from '../../interfaces/mongoose-utils';
-import { Attachment } from '../../models/attachment';
 import { PathAlreadyExistsError } from '../../models/errors';
 import type { PageOperationDocument } from '../../models/page-operation';
 import PageOperation from '../../models/page-operation';
@@ -79,24 +78,17 @@ import PageRedirect from '../../models/page-redirect';
 import type { IRevisionDocument } from '../../models/revision';
 import { Revision } from '../../models/revision';
 import { serializePageSecurely } from '../../models/serializers/page-serializer';
-import ShareLink from '../../models/share-link';
 import Subscription from '../../models/subscription';
 import UserGroupRelation from '../../models/user-group-relation';
 import { V5ConversionError } from '../../models/vo/v5-conversion-error';
 import { divideByType } from '../../util/granted-group';
-import {
-  type ActivityActor,
-  recordCascadeAttachmentRemovals,
-} from '../activity/attachment-removal-snapshot';
+import type { ActivityActor } from '../activity/attachment-removal-snapshot';
 import { configManager } from '../config-manager';
 import type { IPageGrantService } from '../page-grant';
 import { preNotifyService } from '../pre-notify';
 import { getYjsService } from '../yjs';
-import {
-  buildPageIdToPathMap,
-  toAttachmentLikes,
-} from './cascade-attachment-removal-inputs';
 import { BULK_REINDEX_SIZE, LIMIT_FOR_MULTIPLE_PAGE_OP } from './consts';
+import { deleteCompletelyOperation as deleteCompletelyOperationImpl } from './delete-completely-operation';
 import { onSeen } from './events/seen';
 import type { IPageService } from './page-service';
 import { shouldUseV4Process } from './should-use-v4-process';
@@ -2383,64 +2375,14 @@ class PageService implements IPageService {
     return nDeletedNonEmptyPages;
   }
 
-  async deleteCompletelyOperation(
+  deleteCompletelyOperation(
     pageIds,
     pagePaths,
     actor: ActivityActor | null,
   ): Promise<void> {
-    // Delete Attachments, Revisions, Pages and emit delete
-    const Page = mongoose.model<IPage, PageModel>('Page');
-
-    const { attachmentService } = this.crowi;
-    const attachments = await Attachment.find({ page: { $in: pageIds } });
-
-    // Record one ACTION_ATTACHMENT_REMOVE activity per cascaded attachment
-    // BEFORE removeAllAttachments runs: requirement 3.4 demands the snapshot
-    // be frozen before storage deletion, so the cascade records the deletion
-    // "attempt" (design: Error Handling). The recorder isolates per-record
-    // failures internally and never rejects, so page deletion is not blocked.
-    // A null actor means a system operation without an operator
-    // (deleteCompletelyUserHomeBySystem); no operator activity is recorded.
-    if (actor != null) {
-      await recordCascadeAttachmentRemovals(
-        this.crowi.activityService,
-        toAttachmentLikes(attachments),
-        buildPageIdToPathMap(pageIds, pagePaths),
-        actor,
-      );
-    }
-
-    await Promise.all([
-      prisma.$transaction([
-        prisma.comments.deleteMany({
-          where: {
-            pageId: {
-              in: pageIds,
-            },
-            replyToId: {
-              not: null,
-            },
-          },
-        }),
-        prisma.comments.deleteMany({
-          where: {
-            pageId: {
-              in: pageIds,
-            },
-          },
-        }),
-      ]),
-      PageTagRelation.deleteMany({ relatedPage: { $in: pageIds } }),
-      ShareLink.deleteMany({ relatedPage: { $in: pageIds } }),
-      Revision.deleteMany({ pageId: { $in: pageIds } }),
-      Page.deleteMany({ _id: { $in: pageIds } }),
-      PageRedirect.deleteMany({
-        $or: [{ fromPath: { $in: pagePaths } }, { toPath: { $in: pagePaths } }],
-      }),
-      attachmentService.removeAllAttachments(attachments),
-
-      // Leave bookmarks without deleting -- 2024.05.17 Yuki Takei
-    ]);
+    // Thin delegator: the operation body lives in delete-completely-operation.ts
+    // (extracted to keep this file focused). Collaborators are injected via crowi.
+    return deleteCompletelyOperationImpl(this.crowi, pageIds, pagePaths, actor);
   }
 
   // delete multiple pages
